training/random-gen-cnn/test_requirements.txt
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
numpy>=1.21.0
pillow>=9.0.0
opencv-python>=4.5.0
scikit-learn>=1.0.0
torch>=1.11.0
torchvision>=0.12.0
